requirements.txt

apache-beam[gcp]==2.56.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.16.0
